refactor: endpoints foram movidos para o projeto de webAPI #35
