+++ src/Peo.Faturamento.Application/Peo.Faturamento.Application.csproj
@@ -7,7 +7,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Peo.Core.Web\Peo.Core.Web.csproj" />
     <ProjectReference Include="..\Peo.Core\Peo.Core.csproj" />
     <ProjectReference Include="..\Peo.Faturamento.Domain\Peo.Faturamento.Domain.csproj" />
 	<ProjectReference Include="..\Peo.Core.Infra.ServiceBus\Peo.Core.Infra.ServiceBus.csproj" />
