+++ src/Peo.Faturamento.WebApi/Configuration/Dependencies.cs
@@ -4,10 +4,10 @@ using Peo.Core.Infra.ServiceBus.Services;
 using Peo.Core.Interfaces.Services;
 using Peo.Core.Web.Services;
 using Peo.Faturamento.Application.DiConfig;
-using Peo.Faturamento.Application.Endpoints;
 using Peo.Faturamento.Domain.Interfaces.Brokers;
 using Peo.Faturamento.Infra.Data.DiConfig;
 using Peo.Faturamento.Integrations.Paypal.Services;
+using Peo.Faturamento.WebApi.Endpoints;
 
 namespace Peo.Faturamento.WebApi.Configuration
 {
