+++ src/Peo.GestaoAlunos.Application/Peo.GestaoAlunos.Application.csproj
@@ -13,7 +13,6 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\Peo.Core.Infra.ServiceBus\Peo.Core.Infra.ServiceBus.csproj" />
-		<ProjectReference Include="..\Peo.Core.Web\Peo.Core.Web.csproj" />
 		<ProjectReference Include="..\Peo.Core\Peo.Core.csproj" />
 		<ProjectReference Include="..\Peo.GestaoAlunos.Domain\Peo.GestaoAlunos.Domain.csproj" />
 	</ItemGroup>
