+++ src/Peo.GestaoAlunos.WebApi/Configuration/Dependencies.cs
@@ -5,8 +5,8 @@ using Peo.Core.Interfaces.Services;
 using Peo.Core.Web.Services;
 using Peo.GestaoAlunos.Application.Consumers;
 using Peo.GestaoAlunos.Application.DiConfig;
-using Peo.GestaoAlunos.Application.Endpoints;
 using Peo.GestaoAlunos.Infra.Data.DiConfig;
+using Peo.GestaoAlunos.WebApi.Endpoints;
 
 namespace Peo.GestaoAlunos.WebApi.Configuration
 {
