+++ src/Peo.GestaoAlunos.WebApi/Endpoints/EndpointsAluno.cs
@@ -1,12 +1,9 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Routing;
-using Peo.Core.Web.Extensions;
-using Peo.GestaoAlunos.Application.Endpoints.Aula;
-using Peo.GestaoAlunos.Application.Endpoints.Aluno;
-using Peo.GestaoAlunos.Application.Endpoints.Matricula;
+﻿using Peo.Core.Web.Extensions;
+using Peo.GestaoAlunos.WebApi.Endpoints.Aluno;
+using Peo.GestaoAlunos.WebApi.Endpoints.Aula;
+using Peo.GestaoAlunos.WebApi.Endpoints.Matricula;
 
-namespace Peo.GestaoAlunos.Application.Endpoints
+namespace Peo.GestaoAlunos.WebApi.Endpoints
 {
     public static class EndpointsAluno
     {
