+++ src/Peo.GestaoConteudo.WebApi/Configuration/Dependencies.cs
@@ -3,7 +3,6 @@ using Peo.Core.Dtos;
 using Peo.Core.Infra.ServiceBus.Services;
 using Peo.Core.Interfaces.Services;
 using Peo.Core.Web.Services;
-using Peo.GestaoConteudo.Application;
 using Peo.GestaoConteudo.Application.Consumers;
 using Peo.GestaoConteudo.Application.DiConfig;
 using Peo.GestaoConteudo.Infra.Data.DiConfig;
