+++ src/Peo.Identity.WebApi/Endpoints/Responses/LoginResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace Peo.Identity.Application.Endpoints.Responses
+﻿namespace Peo.Identity.WebApi.Endpoints.Responses
 {
     public record LoginResponse(string Token, Guid UserId);
 }
