+++ src/Peo.Identity.WebApi/Endpoints/Responses/RefreshTokenResponse.cs
@@ -1,4 +1,4 @@
-namespace Peo.Identity.Application.Endpoints.Responses
+namespace Peo.Identity.WebApi.Endpoints.Responses
 {
     public record RefreshTokenResponse(string Token, Guid UserId);
 }
